refactor: change decorator folder for api helper
