+++ src/api/dtos/CategoryDTO.ts
@@ -1,5 +1,5 @@
 import CategoryModel from '@models/ICategoryModel';
-import { Min } from '@validaions/Min';
+import { Min } from '../helpers/validations/Min';
 
 export default class CategoryDTO implements Omit<CategoryModel, 'id'> {
   @Min(3, 'Nome da categoria')
