+++ src/api/helpers/validations/isImage.ts
